requirements.txt
google-genai
playwright
flask
gunicorn
google-cloud-secret-manager


